workloadidentity-demo/sampleapp/requirements.txt
azure-identity
azure-storage-blob
uuid